r2dRue-package(r2dRue)	R Documentation

2dRUE model functions

Description

2dRUE is a methodology to make a diagnostic of land condition in a large territory during a given time period. Was developed under the DeSurvey IP.

Details

Package: 	 r2dRue

Type: 	 Package

Version: 	 1.0

Date: 	 2009-12-17

License: 	 GPL

LazyLoad: 	 yes

Author(s)

Gabriel del Barrio, Juan Puigdefabregas, Maria E. Sanjuan & Alberto Ruiz
Maintainer: Alberto Ruiz, <aruiz@eeza.csic.es>

References

Article submitted and under revision.

[Package r2dRue version 1.0 Index]