rasterStack(r2dRue)	R Documentation

Stack bands from raster image files

Description

rasterStack concatenates multiple raster files into a single file. Provides support to several interleaved schemas; BIP (Bands interleaved by pixel), BIL (Band interleaved by line)or BSQ (Bands sequential). The output concatenated file may be saved in Ascii or binary format.

Usage


rasterStack(inFl, outFN, asc = FALSE, zip = FALSE, dec = 3, interleave = "BIP", silent = FALSE)

Arguments

inFl 	 A List with the name of the files to be concatenated

outFN 	 The name of the output file

asc 	 Logical. If TRUE the output is a ASCII file

zip 	 Logical. If TRUE the output is zipped. Only used if asc is TRUE.

dec 	 Integer. If present, the output values are rounded to the specified number of decimal places. Only used if asc is TRUE.

interleave 	 The interleave format. BIP, BIL or BSQ.

silent 	 logical; if TRUE, comment and non-fatal CPL driver errors suppressed

Details

In binary mode (default), the data is stored as 32bit Floating point number. Only the first band of each image is used. This feature is not intended to be of general utility. It is used in internal processes to arrange image data into a format optimum for the process.

Value

An ASCII or binary file.

Author(s)

Alberto Ruiz. aruiz@eeza.csic.es

[Package r2dRue version 1.0 Index]