regStepRaster(r2dRue)	R Documentation

Step Regresion over raster images

Description

This function realices a step regresion

Usage


regStepRaster(ndviFl, timeFl, aridFl, outFl, silent = FALSE, ...)

Arguments

ndviFl 	 Text list (vector) with the file names of the vegetation index grids

timeFl 	 Text list (vector) with the file names of the time series

aridFl 	 Text list (vector) with the file names of the aridity index grids

outFl 	 A list of 7 file names to save the results

silent 	 logical; if TRUE, comment and non-fatal CPL driver errors suppressed

... 	 Any unmatched parameter will be passed to the writeGDAL routine.

Details

ya vere

Value

Returns no value, but save to disk seven image files with the values of

Author(s)

Gabriel del Barrio, Marieta SanJuan & Alberto Ruiz

[Package r2dRue version 1.0 Index]