rgf.create(r2dRue)	R Documentation

Functions to read and write raster group files.

Description

These functions may facilitate working with raster group files.

Usage


rgf.read(inFl)
rgf.create(prefix,suffix='',ini,fin=ini,monthini=1,output)

Arguments

inFl 	 Filename of the Rgf file to be read

prefix 	 Prefix of the output filenames

suffix 	 Sufix of the output filenames

ini 	 Initial year

fin 	 Final year

monthini 	 Initial month (numeric in the 1 to 12 range)

output 	 Filename to be written to disk

Details

rgf.read reads a raster group file and returns a character vector with the read values. rgf.create is useful to generate sequences of file names that contain an indicator of the month and year. Like in Ndvi199001.tif, Ndvi199002.tif ... Ndvi199012.tif, Ndvi199101.tif, Ndvi199102.tif ... Where a prefix (Ndvi) is concatenated to a sequence of years (1990,1991) and months(01,02,...,12) and ended with a suffix (.tif).

Value

Rgf.read returns a character vector with the filenames in the rgf file. Rgf.create returns a character vector. If you supply the output file name parameter, then the list is writen to disc.

Author(s)

Alberto Ruiz, aruiz@eeza.csic.es

Examples


rgf.create('ndvi','.rst',1970,1975,output='ndvi.rgf')
ndvifiles=rgf.read('ndvi.rgf')


[Package r2dRue version 1.0 Index]