solarRad(r2dRue)	R Documentation

Extra-terrestrial solar radiation

Description

solarRad calculate the solar radiation received at the top of the earth's atmosphere on a horizontal surface. This radiation will change during the day and will be different at different latitudes and in different seasons.

Usage


solarRad(img, day = 15)
solarRad12M(img, outFl, ...)

Arguments

img 	 object of class SpatialGridDataFrame that provided a projection class and a bbox attributes. The nature of the data.frame that contains the attribute data, is not relevant. Must be in geographic coordinates.

day 	 The number of the day (from 1 to 365) to calculate the extraterrestrial radiation.

outFl 	 A list of 12 file names to save the results.

... 	 Any unmatched parameter will be passed to the writeGDAL routine.

Details

The output is expressed in cal/cm2*day

Value

solarRad returns a SpatialGridDataFrame class. solarRad12M write on disk 12 images in the format especified by the writeGDAL drivername parameter.

Author(s)

Gabriel del Barrio, Marieta E. SanJuan & Alberto Ruiz

References

Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56

Examples


#not run
#aux=readGDAL('iberia_mask')
#calc the extraterrestial radiation the 15-agost
#Rad=solarRad(aux,227)
#writeGDAL(Rad,'solarrad_15ago')
#or to calculate and write to disk, the radiation of the twelve months
#RadFileNames=paste('rad',month.abb,sep='')
#solarRad12M(img,RadFileNames)


[Package r2dRue version 1.0 Index]